feat: JWT 인증 및 회원 도메인 구현
- Member 필드명 id → userId 변경
- Store 위도/경도 타입 Double → BigDecimal 수정 (Hibernate 6 호환)
- application-dev.yaml 개발 환경 설정 추가 (Redis 제외, JWT/OAuth2 기본값)
- JWT 인증 필터 및 Provider 구현 (JwtAuthenticationFilter, JwtProvider)
- Spring Security 설정 추가 (SecurityConfig, CustomUserDetailsService)
- 회원 도메인 구현 (MemberRepository, MemberService, AuthController, DTO)

+++ src/main/java/com/whale/order/domain/store/entity/Store.java
@@ -8,6 +8,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalTime;
 
 /**
@@ -46,10 +47,10 @@ public class Store extends BaseEntity {
 
     // 지도 API 기반 매장 검색에 사용 (소수점 7자리 = 약 1cm 정밀도)
     @Column(nullable = false, precision = 10, scale = 7)
-    private Double latitude;
+    private BigDecimal latitude;
 
     @Column(nullable = false, precision = 10, scale = 7)
-    private Double longitude;
+    private BigDecimal longitude;
 
     private String phone;
 
@@ -65,7 +66,7 @@ public class Store extends BaseEntity {
 
     @Builder
     public Store(Member owner, String name, String postalCode, String address,
-                 String addressDetail, Double latitude, Double longitude,
+                 String addressDetail, BigDecimal latitude, BigDecimal longitude,
                  String phone, LocalTime openTime, LocalTime closeTime) {
         this.owner = owner;
         this.name = name;
